Move `StorableResources` to `libOPHD`
