+++ appOPHD/MapObjects/Structure.h
@@ -2,12 +2,11 @@
 
 #include "MapObject.h"
 
-#include "../StorableResources.h"
-
 #include <libOPHD/EnumConnectorDir.h>
 #include <libOPHD/EnumDisabledReason.h>
 #include <libOPHD/EnumIdleReason.h>
 #include <libOPHD/EnumStructureID.h>
+#include <libOPHD/StorableResources.h>
 #include <libOPHD/Population/PopulationPool.h>
