+++ appOPHD/MapObjects/Structures/MineFacility.cpp
@@ -4,7 +4,8 @@
 
 #include "../../Constants/Numbers.h"
 #include "../../Constants/Strings.h"
-#include "../../StorableResources.h"
+
+#include <libOPHD/StorableResources.h>
 
 #include <algorithm>
