+++ appOPHD/ProductionCost.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "StorableResources.h"
+#include <libOPHD/StorableResources.h>
 
 
 /**
