+++ appOPHD/States/MapViewStateTurn.cpp
@@ -12,12 +12,12 @@
 
 #include "../Cache.h"
 #include "../MoraleString.h"
-#include "../StorableResources.h"
 #include "../StructureManager.h"
 #include "../Constants/Strings.h"
 
 #include <libOPHD/DirectionOffset.h>
 #include <libOPHD/EnumMoraleIndex.h>
+#include <libOPHD/StorableResources.h>
 
 #include <NAS2D/Utility.h>
 #include <NAS2D/Renderer/Renderer.h>
