+++ appOPHD/appOPHD.vcxproj.filters
@@ -252,9 +252,6 @@
     <ClCompile Include="States\StructureTracker.cpp">
       <Filter>Source Files\States</Filter>
     </ClCompile>
-    <ClCompile Include="StorableResources.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="StructureCatalogue.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
@@ -646,9 +643,6 @@
     </ClInclude>
     <ClInclude Include="States\Wrapper.h">
       <Filter>Header Files\States</Filter>
-    </ClInclude>
-    <ClInclude Include="StorableResources.h">
-      <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="StructureCatalogue.h">
       <Filter>Header Files</Filter>
