+++ libOPHD/StorableResources.cpp
