+++ libOPHD/StorableResources.h
